refactor(components): organize components into separate files
Each component has been moved to its own dedicated file in the components directory.
Enhanced CodeCompanion component with improved state handling and custom spinner frames.
Added proper highlight groups for CodeCompanion spinner states.
